+++ src/client/PlaceView.tsx
@@ -1,24 +1,12 @@
-import NiceModal from "@ebay/nice-modal-react"
-import { map } from "lodash-es"
-import { ArrowLeftRight } from "lucide-react"
-import { useState } from "react"
-import { toast } from "sonner"
 import { api } from "~/utils/api"
 import { Away } from "./Away"
-import { CareButton } from "./CareButton"
-import { MyCatchesModal } from "./MyCatchesModal"
 import { placeTypeIcons } from "./PlaceMarker"
-import { PlaceViewModal } from "./PlaceViewModal"
+import { PlaceViewAirport } from "./PlaceViewAirport"
+import { PlaceViewWildlifeCareCenter } from "./PlaceViewWildlifeCareCenter"
 import { TypeBadge } from "./TypeBadge"
-import { Button } from "./shadcn/ui/button"
-import { Input } from "./shadcn/ui/input"
 import { navigateIcon, runIcon } from "./typeIcons"
-import { useMapFlyTo } from "./useMapRef"
 import { useNavigation } from "./useNavigation"
 import { usePlayer } from "./usePlayer"
-import { DEFAULT_MAP_PITCH, DEFAULT_MAP_ZOOM } from "~/config"
-
-const SHOW_FLUFF = true
 
 export const PlaceView = ({
   placeId,
@@ -48,12 +36,11 @@ export const PlaceView = ({
     <div className="flex flex-col items-center gap-4 pt-12">
       <typeIcon.icon className="w-8 h-8 self-center" />
       <div className="flex flex-col text-center">
-        {SHOW_FLUFF && (
-          <div className="text-sm text-gray-500 italic text-center">
-            Welcome to
-          </div>
-        )}
-        <div>{typeIcon.label}</div>
+        (
+        <div className="text-sm text-gray-500 italic text-center">
+          Welcome to
+        </div>
+        )<div>{typeIcon.label}</div>
         {place.metadata.name && (
           <div className="text-sm opacity-60">{place.metadata.name}</div>
         )}
@@ -85,119 +72,5 @@ export const PlaceView = ({
         />
       </div>
     </div>
-  )
-}
-
-const PlaceViewWildlifeCareCenter = () => {
-  return (
-    <>
-      {SHOW_FLUFF && (
-        <div className="text-sm text-gray-500 italic text-center">
-          This is the perfect place to rest, rejuvenate your team, and prepare
-          for your next battle.
-        </div>
-      )}
-      <div className="flex flex-col w-56">
-        <CareButton />
-      </div>
-      {SHOW_FLUFF && (
-        <div className="text-sm text-gray-500 italic text-center">
-          Don&apos;t forget to manage your team and strategize by swapping
-          moves.
-        </div>
-      )}
-      <div className="flex flex-col w-56">
-        <Button
-          onClick={() => {
-            NiceModal.show(MyCatchesModal)
-          }}
-        >
-          <ArrowLeftRight className="w-4 h-4 mr-2" />
-          <span>Manage my Team</span>
-        </Button>
-      </div>
-      {SHOW_FLUFF && (
-        <div className="text-sm text-gray-500 italic text-center">
-          Remember, every battle is a step closer to becoming the ultimate
-          Wildlife Explorer. Let&apos;s get ready for the next adventure!
-        </div>
-      )}
-    </>
-  )
-}
-const PlaceViewAirport = ({ placeId }: { placeId: string }) => {
-  const { playerId } = usePlayer()
-  const [query, setQuery] = useState("")
-  const { data } = api.place.searchAirports.useQuery(
-    {
-      query,
-      playerId: playerId!,
-    },
-    {
-      enabled: !!playerId && query.length >= 3,
-    }
-  )
-
-  const trpc = api.useContext()
-  const { mutateAsync: fly } = api.place.flyToAirport.useMutation({
-    onSuccess: () => {
-      trpc.invalidate()
-    },
-  })
-
-  const mapFlyTo = useMapFlyTo()
-
-  return (
-    <>
-      <div>
-        <Input
-          value={query}
-          onChange={(e) => setQuery(e.target.value)}
-          className="w-96"
-        />
-      </div>
-      <div className="flex flex-col gap-1 w-96">
-        {map(data, (airport) => (
-          <button
-            key={airport.code}
-            className="flex flex-row gap-2 rounded p-2 bg-gray-100 hover:bg-gray-200 text-xs items-center text-left"
-            onClick={async () => {
-              if (!confirm(`Fly to ${airport.name}?`)) return
-              if (!playerId) return
-              const promise = fly({
-                playerId,
-                destinationCode: airport.code,
-                placeId,
-              })
-              toast.promise(promise, {
-                loading: "Flying...",
-                success: "You have arrived!",
-                error: (err: any) => err?.message || "Something went wrong",
-              })
-              try {
-                await promise
-                NiceModal.hide(PlaceViewModal)
-                mapFlyTo({
-                  center: airport,
-                  pitch: DEFAULT_MAP_PITCH,
-                  zoom: DEFAULT_MAP_ZOOM,
-                })
-              } catch (e) {
-                return
-              }
-            }}
-          >
-            <div className="flex flex-col overflow-hidden">
-              <div className="truncate">{airport.name}</div>
-              <div className="truncate">
-                {airport.city}, {airport.country}
-              </div>
-            </div>
-            <div className="flex-1" />
-            <Away location={airport} />
-          </button>
-        ))}
-      </div>
-    </>
   )
 }
